+++ CHANGES.rst
@@ -3,7 +3,7 @@
 Version 2.1.0
 -------------
 
-Unreleased
+Released 2022-03-28
 
 -   Drop support for Python 3.6. :pr:`4335`
 -   Update Click dependency to >= 8.0. :pr:`4008`